+++ Create-Release.ps1
@@ -5,17 +5,16 @@ if (Test-Path .\Release) {
 }
 
 mkdir .\Release
-mkdir .\Release\Portable
-mkdir .\Release\Publish
+mkdir .\Release\EsperaPortable
 
 # Build the portable version
 Write-Host "Building Portable Version"
 & "$MSBuildLocation\MSBuild.exe" /t:Rebuild /p:Configuration=Release /p:Platform="x86" /v:quiet ".\Espera\Espera.sln"
 
-cp ".\Espera\Espera.View\bin\Release\*.dll" ".\Release\Portable\"
-cp ".\Espera\Espera.View\bin\Release\Espera.exe" ".\Release\Portable\"
-cp ".\Espera\Espera.View\bin\Release\Espera.exe.config" ".\Release\Portable\"
-cp ".\Changelog.md" ".\Release\Portable\Changelog.txt"
+cp ".\Espera\Espera.View\bin\Release\*.dll" ".\Release\EsperaPortable\"
+cp ".\Espera\Espera.View\bin\Release\Espera.exe" ".\Release\EsperaPortable\"
+cp ".\Espera\Espera.View\bin\Release\Espera.exe.config" ".\Release\EsperaPortable\"
+cp ".\Changelog.md" ".\Release\EsperaPortable\Changelog.txt"
 
 # Build the ClickOnce version
 Write-Host "Building ClickOnce version"
@@ -26,5 +25,3 @@ Write-Host "Building ClickOnce version"
 
 cp -r ".\Espera\Espera.View\bin\Release\app.publish\" ".\Release\"
 Rename-Item ".\Release\app.publish\setup.exe" "EsperaSetup.exe"
-
-